add in more things to pmap_reset temporarily, since it breaks some stuff if these aren't reset
will probably stop with the c++ for now, and work on version 24.1 of my mod. Having c++, and the redone elements was necessary for continuing some things I wanted to do. I will make a third branch before continuing, and use this second branch in the future until the third one is stable