rewrite a lot of stuff to use a new Tool class. Also redo a lot of things with Menus, redo a little with drawing, and start on a Brush class.
Everything tested and working, including the fav. menu, the old menu, element search, find tool, and some other things. Some bugs were fixed with the fav menu (and more created), and a lot fixed with other random things.

Even though it all works it's sort of in an unfinished state. I got rid of a lot of the TODO's before commiting though. I'll probably continue making things work in an unfinshed state, fixing the bugs, and then releasing new versions in the future. Wall ID's and similar need to be redone to start at one first.